securid: Improve handling of device-bound tokens
TokenConverter allows a DeviceSerialNumber that is longer than the
40-byte (smartphone) or 32-byte (feature phone) limit, so we should
accept ctf strings that were generated from an sdtid file containing
such a device ID, and perform the appropriate truncation inside the
library.

TokenConverter also automatically converts the DeviceSerialNumber to
uppercase, even if it was lowercase in the sdtid file.  We will have the
library perform the same conversion.

fixes #3

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>