sdtid: Enhance XML utility functions
The lookup_* functions will now perform a recursive lookup in order to
handle the TokenAttributes node.  Fortunately every node in this document
format has a unique name so there are no conflicts.

Add a replace_string() function used to build new sdtid documents.

And add a node_present() function used to determine whether a field is
present in a template document.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>