build-debian.sh: Automate PPA builds
If this script is run on an Ubuntu system, it will generate an appropriate
changelog entry and name the new version something like
"0.5~20140313-1ubuntu1~precise".  The first tilde will cause it to be
superseded by "0.5"[1].  The resultant *.changes file can be directly
uploaded to PPA/Launchpad.

Debian sid behavior is unchanged, as we will continue to use sid for
official uploads.

[1] https://www.debian.org/doc/debian-policy/ch-controlfields.html#s-f-Version

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>