securid: v3: Disable has_enc_seed
v3 tokens do have an encrypted seed, but it is part of a larger encrypted
blob so it doesn't fit neatly into the enc_seed[] field.  Since this
flag is only checked by securid_token_info(), we'll disable it for now
so that "stoken show --seed" doesn't show all zeroes for the encrypted
seed.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>