library: Force stoken_encrypt_seed() to create v2 tokens
If a v3 token is imported, then stoken_encrypt_seed() is called, it will
try to create a v3 token.  This is a problem for two reasons: first, it
is unexpected behavior from the caller's perspective; second, it causes
a heap overflow because only (MAX_TOKEN_CHARS + 1) bytes were allocated
to store the string.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>