java: Relicense the JNI wrapper under BSD-3-clause
In some circumstances, such as an Android app, it is difficult to provide
discrete .class files that comply with section 6(a) of the LGPL, and
dynamic linking under section 6(b) is not feasible.  So we'll keep the
main library (C code) under LGPLv2.1 to ensure that end users can still
replace our .so file with a modified version, but license the Java wrapper
file more permissively.

Fixes #5

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>