gui: Rewrite code to use GTK+3 and glade
Define most windows and dialogs in XML instead of C.  Add next tokencode
and other info to the "detailed" stoken-gui window, making it look more
like Easy Token on Android.  Use a custom progressbar widget to avoid GTK+
quirks/decorations.

Also, this splits the password and PIN prompts, as v3 tokens need to be
decrypted with the password before we know whether a PIN is needed.

Closes #6.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>