Bypass securid_decrypt_seed() sanity checks when testing device IDs
securid_decrypt_seed() will abort immediately if the token requires a
password but a password isn't provided.  This is *usually* what we want,
but when we're just trying to guess the bound device ID it causes
the function to return prematurely.  In this case we will pass in "." as
the password so that we can see whether the devid hash check returns
ERR_BAD_DEVID or (anything else).

This is still safe on passwordless tokens, because in that case the logic
just ignores the supplied password.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>