sdtid: Fix corner case in MAC calculation
In order to calculate the Header and Token MACs, we first convert each
string from the XML file into a key/value string.  For example, this
input:

  <TKN>
    <AddPIN>1</AddPIN>
  </TKN>

becomes: "TKN.AddPIN 1\n".  Then a specific number of zero bytes are
appended, based on the starting offset in the AES block.  This is all
kind of silly, but our MAC won't match RSA's MAC if we don't follow the
same convention.

The problem is that RSA's implementation has a quirk: if the string plus
padding bytes all fits in a single AES block, then it needs to be written
out starting at offset 0 in the block.  The lack of this extra logic
caused stoken to incorrectly reject a small number of valid sdtid files.

So, let's add a check for this, and a test case to make sure it doesn't
break later.

fixes #21

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>